Revert "Feature/get my bookmarks #22"
