+++ backend/backend/common/dao/resources.py
@@ -8,23 +8,29 @@ from backend.common import schema
 def get_resources(db, offset=0, limit=100, search=""):
     s: Final[str] = f"%{search}%".lower()
     try:
-        query = db.query(models.Resource)
         if search:
-            query.filter(
+            resources: List[models.Resource] = db.query(
+                models.Resource
+            ).filter(
                 or_(
                     models.resource.name.like(s),
                     models.Resource.description.like(s),
                 )
+            ).offset(offset).limit(limit).all()
+
+            return schema.DatabaseResult(
+                success=True,
+                data=resources
             )
+        else:
+            resources = db.query(
+                models.Resource
+            ).offset(offset).limit(limit).all()
 
-        resources: List[models.Resource] = query.offset(
-            offset
-        ).limit(limit).all()
-
-        return schema.DatabaseResult(
-            success=True,
-            data=resources
-        )
+            return schema.DatabaseResult(
+                success=True,
+                data=resources
+            )
     except exc.SQLAlchemyError as e:
         return schema.DatabaseResult(
             success=False,
